ai-planning-coworker/requirements.txt
anthropic
fastapi
uvicorn
python-dotenv
httpx
aiofiles
python-multipart
